+++ lib/easy_video.dart
@@ -1,9 +1,3 @@
 library easy_video;
 
 export 'src/easy_video_widget.dart';
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}