+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Easy Video Player
+- Support video mp4, m3u8,...
+- Support multiplatform: ios, android, web